binance/trade_app/requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.10.4
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
bitarray==2.7.1
ccxt==2.7.78
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.1
cytoolz==0.12.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
fastapi==0.91.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
hexbytes==0.3.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
Mako==1.2.4
MarkupSafe==2.1.2
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
parsimonious==0.8.1
pkgutil-resolve-name==1.3.10
protobuf==3.19.5
psycopg2-binary==2.9.6
pycares==4.3.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.4
pyrsistent==0.19.3
python-dotenv==0.21.1
requests==2.28.2
rlp==2.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
starlette==0.24.0
toolz==0.12.0
typing-extensions==4.4.0
urllib3==1.26.14
uuid==1.30
uvicorn==0.20.0
varint==1.0.2
web3==5.31.3
websocket-client==1.5.1
websockets==9.1
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.13.0
